requirements.txt
colorthief==0.2.1
Pillow==9.4.0
requests==2.28.2
webcolors==1.12
